+++ src/appellate/utils.js
@@ -457,13 +457,26 @@ let APPELLATE = {
     );
   },
 
-  // Adds the vue data attributes to the session storage
-  storeVueDataInSession: () => {
+  // Adds data attributes to the session storage
+  storeMetaDataInSession: () => {
     return new Promise((resolve, reject) =>
-      chrome.runtime.sendMessage({ message: 'getVueData' }, (res) => {
+      chrome.runtime.sendMessage({ message: 'getMetaData' }, (res) => {
         if (res == null) reject('Response cannot be null');
         resolve(res);
       })
+    );
+  },
+
+  // Sends message to background to fetch document data from download modal
+  storeDocumentDataFromDownloadModal: () => {
+    return new Promise((resolve, reject) =>
+      chrome.runtime.sendMessage(
+        { message: 'getDocumentDataFromDownloadModal' },
+        (res) => {
+          if (res == null) reject('Response cannot be null');
+          resolve(res);
+        }
+      )
     );
   },
 
@@ -518,7 +531,7 @@ let APPELLATE = {
   },
 
   // Creates a div with a spinner and a loading message
-  createsLoadingMessage: (downloadData) => {
+  createsLoadingMessage: (caseNumber) => {
     let loadingTextDiv = document.createElement('div');
     loadingTextDiv.classList.add('box', 'mt-2');
 
@@ -529,9 +542,13 @@ let APPELLATE = {
     let spanText = document.createElement('span');
     spanText.innerHTML =
       'Download in progress for case number ' +
-      `${downloadData.caseSummary.caseDetails.caseNumber}`;
+      `${caseNumber}`;
+
+    // Add a non-breaking space between spinner and text
+    const space = document.createTextNode('\u00A0');
 
     loadingTextElement.appendChild(spinner);
+    loadingTextElement.appendChild(space);
     loadingTextElement.appendChild(spanText);
     loadingTextDiv.appendChild(loadingTextElement);
     return loadingTextDiv;